Refactor history feature package structure
