+++ feature/history/src/main/kotlin/dev/nevack/unitconverter/history/di/HistoryModule.kt
@@ -1,4 +1,4 @@
-package dev.nevack.unitconverter.history
+package dev.nevack.unitconverter.history.di
 
 import android.content.Context
 import androidx.room.Room
@@ -9,8 +9,10 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import dev.nevack.unitconverter.history.db.HistoryDao
-import dev.nevack.unitconverter.history.db.HistoryDatabase
+import dev.nevack.unitconverter.history.HistoryRepository
+import dev.nevack.unitconverter.history.data.RoomHistoryRepository
+import dev.nevack.unitconverter.history.data.db.HistoryDao
+import dev.nevack.unitconverter.history.data.db.HistoryDatabase
 
 @Module
 @InstallIn(SingletonComponent::class)
