+++ feature/history/src/main/kotlin/dev/nevack/unitconverter/history/usecase/ClearHistoryUseCase.kt
@@ -1,5 +1,6 @@
-package dev.nevack.unitconverter.history
+package dev.nevack.unitconverter.history.usecase
 
+import dev.nevack.unitconverter.history.HistoryRepository
 import javax.inject.Inject
 
 class ClearHistoryUseCase
