+++ feature/history/src/main/kotlin/dev/nevack/unitconverter/history/usecase/GetHistoryUseCase.kt
@@ -1,5 +1,7 @@
-package dev.nevack.unitconverter.history
+package dev.nevack.unitconverter.history.usecase
 
+import dev.nevack.unitconverter.history.HistoryRecord
+import dev.nevack.unitconverter.history.HistoryRepository
 import javax.inject.Inject
 
 class GetHistoryUseCase
